tests/wpt/css-tests/cssom-view-1_dev/html/ttwf-js-cssomview-getclientrects-length.htm

<!DOCTYPE html>
<html><head>
    <title>CSSOM View API Test: the length of getClientRects</title>
    <link href="mailto:zhaolp0419@gmail.com" rel="author" title="simplezeroec">
    <link href="https://www.w3.org/TR/cssom-view/#dom-element-getclientrects" rel="help">
    <meta content="dom" name="flags">
    <meta content="getClientRects will return rects of the correct number" name="assert">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

</head>
<body>
	<div id="rectcontainer">
		<button id="testBtn">Test</button>
		<a id="testLink">Test Link</a>
	</div>
    <div id="log"></div>
    <script>

        test(function(){assert_equals(document.getElementById("rectcontainer").getClientRects().length,1,"1 ClientRects should be returned");},"assert_length_of_getClientRects_from_Parent");
		test(function(){assert_equals(document.getElementById("testBtn").getClientRects().length,1,"1 ClientRects should be returned");},"assert_length_of_getClientRects_from_Button");
		test(function(){assert_equals(document.getElementById("testLink").getClientRects().length,1,"1 ClientRects should be returned");},"assert_length_of_getClientRects_from_Link");


    </script>


</body></html>
